+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 These are changes made in the repository to the official version 3.09b.
 
+- 2021-04-11: Removed deprecated `pcap_lookupdev` by `pcap_findalldevs`. Added
+  a new function `find_interface_unix` to search of device that can be opened.
 - 2021-04-11: Reorganized the source tree.
 - 2021-04-11: Added cmake build system.